update isometric angle

+++ src/vector.c
@@ -6,13 +6,13 @@
 /*   By: nde-sant <nde-sant@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/15 13:25:44 by nde-sant          #+#    #+#             */
-/*   Updated: 2025/11/15 13:29:05 by nde-sant         ###   ########.fr       */
+/*   Updated: 2025/11/24 12:40:44 by nde-sant         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h"
 
-void	set_vector(t_vector *vector, int x, int y, int z)
+void	set_vector(t_vector *vector, double x, double y, double z)
 {
 	vector->x = x;
 	vector->y = y;
